Enable the Graphical User Interface again.
You see, this is a very important commit. In the last commit, I turned the GUI off. But, when i went to push it, I forgot to turn it back on. So, people who don't know anything about programming will have a hard time using it. Re-enabling the GUI fixes this issue, because then there will be a user friendly interface by default.